+++ MSGraph/functions/exchange/mail/message/Add-MgaMailMessageForward.ps1
@@ -135,7 +135,7 @@
                 switch ($PSCmdlet.ParameterSetName) {
                     'Default' { $invokeParam.Add("Field", "messages/$($messageItem.Id)/createForward") }
                     'DirectReply' { $invokeParam.Add("Field", "messages/$($messageItem.Id)/forward") }
-                    Default { stop-PSFMessage -Message "Unhandled parameter set. ($($PSCmdlet.ParameterSetName)) Developer mistake." -EnableException $true -Category "ParameterSetHandling" -FunctionName $MyInvocation.MyCommand }
+                    Default { Stop-PSFFunction -Message "Unhandled parameter set. ($($PSCmdlet.ParameterSetName)) Developer mistake." -EnableException $true -Category MetadataError -FunctionName $MyInvocation.MyCommand }
                 }
 
                 $output = Invoke-MgaRestMethodPost @invokeParam