+++ library/MSGraph/MSGraph/Exchange/ExchangeCommonEnums.cs
@@ -1,0 +1,4 @@
+﻿namespace MSGraph.Exchange
+{
+
+}