+++ library/MSGraph/MSGraph/Exchange/MailboxSetting/MailboxSettingEnums.cs
@@ -1,0 +1,44 @@
+﻿namespace MSGraph.Exchange.MailboxSetting
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public enum AutomaticRepliesStatus
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        Disabled,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        AlwaysEnabled,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        Scheduled
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    public enum ExternalAudienceScope
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        None,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        ContactsOnly,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        All
+    }
+}